Rename chrono_win_patched to chrono_with_profiling_clock, add profiling_clock typedef to boost_chrono header
